Almost ready...
I added that useless editor that allows you to add nodes
definitions and types.
The nodes definitions will be still be useful... The types,
not really. They're not going to change that much, so,
yeah...

Anyway, it's now possible to save and load, though I'll have
to add a real FileSelector at the end.

What remains now are :
* Constant definitions
* Variable definitions
* Dynamic variables definitions

Dynamic variables are actually easier to define that variables,
the latter require a combination of Register, LogixReference,
ReferenceNode<IValue> and, by default, the game seems to define
them wherever it wants, so I guess I'll go for DynVars first,
then simple registers afterwards.

Constants only require a specific 'Input' node. You cannot write
to them, so they should be simpler to implement.

I'll still try to go for a quick release before these 3 features,
even if they're important, because I'll like to gauge the
reception of such editors.
I mean, I mainly built this to write a JSON parser on NeosVR
and import the emotions created by my Quick Avatar Setup tool.

Signed-off-by: Voyage <voyage@miouyouyou.fr>