Almost ready for the first alpha
Still need to add proper file selectors (instead of Save/Load buttons),
check if this works well in webbrowsers, (I'll certainly have to add
Websocket RELAY support...) and then kind of 'clean up' the BIG MESS
that TestingGraph.gd is.

Still, this is TESTED AND WORKING WITH NEOSVR !

YAAAAY ! WOOHOO !

So, let's clean things a little and get ready for the first release.

Signed-off-by: Voyage <voyage@miouyouyou.fr>