Starting to clean things up and testing ideas for generics
So, the next step now is, getting almost ALL the nodes available
in Neos, in order to make the tool useful for making any kind of
LogiX program.

To support all the nodes, we need :
* A real support for generics
Generics come in three forms :
  * One form require the user to actually input the specific types.
    This is true for *DynamicVariable for example.
  * One form specialize itself depending on the input connections.
  * And another form do the same, but is not really a generic

For example, NotNullNode is an actual generic, which take the
input connection type to specialize itself. So if you pass a
Slot, it becomes a NotNullNode<FrooxEngine.Slot>.

The operator Add (+) kind of do the same, but instead of morphing
into Add<System.Single> when you connect a float, it morphs into
Add_Float. And if you connect a color to the 'B' slot, it becomes
Add_Float_Color... Yeah...

I will deal with the two first elements, and see if I deal with
last one now or for the next release. Dealing with the third case
doesn't seem very difficult, but this will require a good dictionary
of all the appropriate combinations, and that also require to
reconfigure the slots with specific types that accept connections
from only the various supported types... Which is the main issue
actually.
I could go for a dummy slot and deny the connection if it's not
possible, just like Neos does it, though.
From what I see, these combinations only target primitive types,
so there might be a way to tackle this nicely.

Meanwhile, I started to split things up. The nodes can now be
nicely themed using the Godot Editor, just like it should be.
The new design is not necessarily better in every regard, in
terms of integration with Godot, it is !

Signed-off-by: Voyage <voyage@miouyouyou.fr>