A glimpse of the new generic system
So I now have an idea of how I can implent this for the
nodes definitions and export... What remains is the update
of the ports and connections when updating the type.

Also moving back the devel branch to main, since the
first release is already tagged correctly.
Main branch is main branch. If you want the code of the
first release, pull from v0alpha1.

Signed-off-by: Voyage <voyage@miouyouyou.fr>