Almost done with the generic with menus
THe drawbacks from the first release rush is that the architecture
of the program is slowing down updates.
The whole editor required too much useless code and will probably
be removed from the main app on the next version.
It should come back as a theme editor in the future, but the whole
"let's add nodes by hand, by providing a C# class name and editing
slots, one by one" part will be removed, since most of the nodes
will soon be added through a script, and there shouldn't be any
need for such editor.
90% of the users don't care, the remaining 10% know how to edit a
JSON file.

Still, I'm starting to have a working version of
WriteDynamicVariable`1 settable through a menu. There's still bugs
here and there, but it's starting to get shape.
I'll still try to get the generic through i/o connections and the
pseudo generic working almost the same way before fixing everything,
so that I have an idea of what to factorize and cleanup.

Then, if everything seems fine, I'll start to fix the popup menu,
using a tree based menu for the node selection, and then add all
the known logix nodes in NeosVR.

The last part will be to support writing to register nodes correctly,
and we should be done for alpha 2.

Signed-off-by: Voyage <voyage@miouyouyou.fr>