Splitting nodes and types definitions
The types definitions are more of a 'theme' definition than
anything else.

This is in preparation for the nodes list generator.

A few things are still remaining, and I don't know how to
tackle this nicely though :
* Write nodes and their slightly weird generic behaviour
  will surely require a specific class to implement them
  correctly.
* Multi nodes will also require some work, and might require
  a few changes to the protocol.
* Inputs needs to have their own look, based on which input
  you provide.

There's a also a few odd nodes, like "Relay", which will
require a few look at the Wiki to understand their usage.

Signed-off-by: Voyage <voyage@miouyouyou.fr>