Trying to advance with the mess
This whole grid system needs a good rewrite. I've tested
multiple Godot projects using this Grid system (Material
Maker, Concept Graph) and they all have the same broken
interaction issues.

Also my current program design tends to be horrible, and
the weird Grid API is NOT helping ONE BIT.

So I'm still trying to get with this "mess" for the Alpha
2, and at least registers, Write nodes, and nice input
nodes working.
Then for the alpha 3, I'll try to rework this grid "thing"
and make it work more nicely... Or see if there's another
grid system on other MIT-licensed engines out there, that
could be used for this task.

Signed-off-by: Voyage <voyage@miouyouyou.fr>